+++ airdrop/project/SPECIFICATION.md
@@ -1,16 +1,18 @@
 Table of Contents
 - [Overview](#overview)
 - [Use Cases](#use-cases)
-  - [Actions that users are able to perform](#actions-that-users-are-able-to-perform)
-    - [Airdrop Core Functionality](#airdrop-core-functionality)
+  - [Airdrop](#airdrop)
+    - [Core Functionality](#core-functionality)
       - [`constructor()`](#constructor)
       - [`claim()`](#claim)
-    - [Airdrop State Checks](#airdrop-state-checks)
+    - [State Checks](#state-checks)
       - [`claim_data()`](#claim_data)
       - [`end_block()`](#end_block)
       - [`merkle_root()`](#merkle_root)
-    - [Simple Asset Core Functionality](#simple-asset-core-functionality)
+  - [Simple Asset](#simple-asset)
+    - [Core Functionality](#core-functionality-1)
       - [`constructor()`](#constructor-1)
+      - [`mint_to()`](#mint_to)
   - [Sequence Diagram](#sequence-diagram)
 
 # Overview
@@ -27,11 +29,9 @@ This section contains general information about the functionality of the applica
 
 If you are interested in a functional overview then this is the section for you.
 
-## Actions that users are able to perform
+## Airdrop
 
-This sub-section details what a user is able to do e.g. click a button and "x, y, z" happens.
-
-### Airdrop Core Functionality
+### Core Functionality
 
 #### `constructor()`
 
@@ -46,7 +46,7 @@ This sub-section details what a user is able to do e.g. click a button and "x, y
     1. They have provided a valid Merkle Proof
     2. The deadline has not passed
 
-### Airdrop State Checks
+### State Checks
 
 #### `claim_data()`
 
@@ -62,13 +62,19 @@ This sub-section details what a user is able to do e.g. click a button and "x, y
 
 1. Returns the stored merkle root used for claim validation
 
-### Simple Asset Core Functionality
+## Simple Asset 
+
+### Core Functionality
 
 #### `constructor()`
 
 1. Allows the owner to begin minting the asset and requires
     1. An asset supply be given
     2. A user that will be authorized to mint
+
+#### `mint_to()`
+
+TODO
 
 ## Sequence Diagram
