+++ airdrop/project/contracts/distributor-contract/tests/functions/core/claim.rs
@@ -1,13 +1,14 @@
 use crate::utils::{
-    airdrop_distributor_abi_calls::{airdrop_constructor, claim, claim_data},
-    simple_asset_abi_calls::asset_constructor,
-    test_helpers::{build_tree, build_tree_manual, defaults, leaves_with_depth, setup},
+    interface::core::{airdrop_constructor, asset_constructor, claim},
+    setup::{build_tree, build_tree_manual, defaults, setup},
 };
-use fuels::{tx::AssetId, types::Identity};
+use fuels::tx::AssetId;
 
 mod success {
 
     use super::*;
+    use crate::utils::{interface::info::claim_data, setup::leaves_with_depth};
+    use fuels::types::Identity;
 
     // NOTE: This test is ignored as it uses the Fuel-Merkle crate. There is currently an
     // incompatability with the Fuel-Merkle crate and the Sway-Libs Merkle Proof library.
