+++ airdrop/project/contracts/distributor-contract/tests/functions/core/mod.rs
@@ -1,0 +1,2 @@
+mod claim;
+mod constructor;
