+++ airdrop/project/contracts/distributor-contract/tests/functions/info/end_block.rs
@@ -1,12 +1,10 @@
-use crate::utils::{
-    airdrop_distributor_abi_calls::{airdrop_constructor, end_block},
-    test_helpers::{defaults, setup},
-};
-use fuels::types::Bits256;
-
 mod success {
 
-    use super::*;
+    use crate::utils::{
+        interface::{core::airdrop_constructor, info::end_block},
+        setup::{defaults, setup},
+    };
+    use fuels::types::Bits256;
 
     #[tokio::test]
     async fn returns_end_block() {
