+++ airdrop/project/contracts/distributor-contract/tests/functions/info/mod.rs
@@ -1,0 +1,3 @@
+mod claim_data;
+mod end_block;
+mod merkle_root;
