+++ airdrop/project/contracts/distributor-contract/tests/functions/mod.rs
@@ -1,5 +1,2 @@
-mod claim;
-mod claim_data;
-mod constructor;
-mod end_block;
-mod merkle_root;
+mod core;
+mod info;
