+++ airdrop/project/contracts/distributor-contract/tests/utils/interface/mod.rs
@@ -1,0 +1,2 @@
+pub mod core;
+pub mod info;
